+++ ants/__init__.py
@@ -3,8 +3,10 @@ __version__ = '0.5.3'
 
 from .core import *
 from .utils import *
+from .metrics import *
+from .transforms import *
 from .segmentation import *
-from .registration import *
+#from .registration import *
 from .learn import *
 from .viz import *
