+++ ants/registration/__init__.py
@@ -7,15 +7,15 @@ from .create_jacobian_determinant_image import deformation_gradient
 from .create_warped_grid import create_warped_grid
 from .fsl2antstransform import fsl2antstransform
 from .make_points_image import make_points_image
-from .metrics import image_mutual_information
-from .reflect_image import reflect_image
-from .reorient_image import (get_orientation,
+from ..metrics.metrics import image_mutual_information
+from ..ops.reflect_image import reflect_image
+from ..ops.reorient_image import (get_orientation,
            reorient_image2,
            get_possible_orientations,
            get_center_of_mass)
-from .resample_image import (resample_image,
+from ..ops.resample_image import (resample_image,
            resample_image_to_target)
-from .symmetrize_image import symmetrize_image
+from ..ops.symmetrize_image import symmetrize_image
 from .build_template import build_template
 from .landmark_transforms import (fit_transform_to_paired_points, fit_time_varying_transform_to_point_sets)
