+++ ants/registration/affine_initializer.py
@@ -6,6 +6,7 @@ from tempfile import mktemp
 
 from .. import utils
 
+
 def affine_initializer(fixed_image, moving_image, search_factor=20,
                         radian_fraction=0.1, use_principal_axis=False, 
                         local_search_iterations=10, mask=None, txfn=None ):
