+++ ants/registration/apply_transforms.py
@@ -4,9 +4,11 @@ __all__ = ['apply_transforms','apply_transforms_to_points']
 
 import os
 
+
 from .. import core
 from ..core import ants_image as iio
 from .. import utils
+
 
 def apply_transforms(fixed, moving, transformlist,
                      interpolator='linear', imagetype=0,
