+++ ants/utils/__init__.py
@@ -1,12 +1,12 @@
-from .add_noise_to_image import add_noise_to_image
-from .bias_correction import (n3_bias_field_correction, n3_bias_field_correction2, n4_bias_field_correction, abp_n4)
+from ..ops.add_noise_to_image import add_noise_to_image
+from ..ops.bias_correction import (n3_bias_field_correction, n3_bias_field_correction2, n4_bias_field_correction, abp_n4)
 from .channels import merge_channels, split_channels
 from .compose_displacement_fields import compose_displacement_fields
 from .convert_nibabel import (to_nibabel, from_nibabel, nifti_to_ants)
-from .crop_image import (crop_image, 
+from ..ops.crop_image import (crop_image, 
            crop_indices,
            decrop_image)
-from .denoise_image import denoise_image
+from ..ops.denoise_image import denoise_image
 from .fit_bspline_object_to_scattered_data import fit_bspline_object_to_scattered_data
 from .fit_bspline_displacement_field import fit_bspline_displacement_field
 from .fit_thin_plate_spline_displacement_field import fit_thin_plate_spline_displacement_field
@@ -57,13 +57,13 @@ from .label_image_centroids import label_image_centroids
 from .label_overlap_measures import label_overlap_measures
 from .label_stats import label_stats
 from .labels_to_matrix import labels_to_matrix
-from .mask_image import mask_image
+from ..ops.mask_image import mask_image
 from .mni2tal import mni2tal
 from .morphology import morphology
 from .multi_label_morphology import multi_label_morphology
 from .ndimage_to_list import (ndimage_to_list,
            list_to_ndimage)
-from .pad_image import pad_image
+from ..ops.pad_image import pad_image
 from .process_args import (
     get_pointer_string,
     short_ptype,
@@ -81,10 +81,10 @@ from .quantile import (ilr,
            bandpass_filter_matrix )
 from .scalar_rgb_vector import (rgb_to_vector, vector_to_rgb, scalar_to_rgb)
 from .simulate_displacement_field import simulate_displacement_field
-from .slice_image import slice_image
-from .smooth_image import smooth_image
-from .threshold_image import threshold_image
+from ..ops.slice_image import slice_image
+from ..ops.smooth_image import smooth_image
+from ..ops.threshold_image import threshold_image
 from .weingarten_image_curvature import weingarten_image_curvature
-from .average_transform import average_affine_transform, average_affine_transform_no_rigid
+from ..transforms.average_transform import average_affine_transform, average_affine_transform_no_rigid
 from .averaging import average_images
