+++ ants/utils/label_image_centroids.py
@@ -4,7 +4,7 @@ __all__ = ['label_image_centroids']
 
 import numpy as np
 
-from ..core import ants_transform as tio
+from ..transforms import ants_transform as tio
 
 
 def label_image_centroids(image, physical=False, convex=True, verbose=False):
