+++ ants/utils/ndimage_to_list.py
@@ -1,13 +1,15 @@
 __all__ = ['ndimage_to_list',
            'list_to_ndimage']
 
+
 import numpy as np
-# import itk as itk
+
+from ants.decorators import image_method
 from ..core import ants_image as iio
 from ..core import ants_image_io as iio2
 from .. import utils
 
-
+@image_method
 def list_to_ndimage( image, image_list ):
     """
     Merge list of multiple scalar ANTsImage types of dimension into one
@@ -63,7 +65,7 @@ def list_to_ndimage( image, image_list ):
 
 
 
-
+@image_method
 def ndimage_to_list(image):
     """
     Split a n dimensional ANTsImage into a list
