+++ ants/utils/simulate_displacement_field.py
@@ -2,10 +2,12 @@ __all__ = ["simulate_displacement_field"]
 
 import numpy as np
 
+from ants.decorators import image_method
 from ..core import ants_image as iio, ants_image_io as iio2
 from .. import utils
 
 
+@image_method
 def simulate_displacement_field(domain_image, 
                                 field_type="bspline", 
                                 number_of_random_points=1000,
