+++ ants/viz/plot.py
@@ -24,13 +24,14 @@ from mpl_toolkits.axes_grid1.inset_locator import inset_axes
 
 import numpy as np
 
+from ants.decorators import image_method
 from .. import registration as reg
 from ..core import ants_image as iio
 from ..core import ants_image_io as iio2
-from ..core import ants_transform as tio
-from ..core import ants_transform_io as tio2
+from ..transforms import ants_transform as tio
+from ..transforms import ants_transform_io as tio2
 
-
+@image_method
 def plot(
     image,
     overlay=None,
