+++ ants/viz/plot_ortho.py
@@ -27,8 +27,8 @@ import numpy as np
 from .. import registration as reg
 from ..core import ants_image as iio
 from ..core import ants_image_io as iio2
-from ..core import ants_transform as tio
-from ..core import ants_transform_io as tio2
+from ..transforms import ants_transform as tio
+from ..transforms import ants_transform_io as tio2
 
 def plot_ortho(
     image,
